requirements.txt
Django>=4.2,<5.0
django-filter>=23,<24
openpyxl>=3.1.0
reportlab>=4.0.0

# Produção (Docker / Gunicorn / PostgreSQL)
gunicorn>=21.0,<23
psycopg2-binary>=2.9.9
whitenoise>=6.6,<7
